requirements.txt
pytest
selene >= 2.0.0b4
allure-pytest